+++ CHANGES.md
@@ -8,6 +8,12 @@ Version 0.1.1
 
 To be released.
 
+ -  Fixed a bug where the configured sinks and filters were reset after
+    some inactivity.  This was caused by garbage collection of the
+    `Logger` instances.  The `Logger` instances are now kept alive by
+    an internal set of strong references until explicitly `reset()` is
+    called.
+
 
 Version 0.1.0
 -------------